// src/app/models/common-entities.ts
export interface FacilityEntiry {
  facility_id: string;
  facility_name: string;
}

export interface OptionEntity {
  name: string;
}

export interface DepartmentEntity {
    department_id: string;
    department_name: string;
}

export interface ClinicEntity {
    clinic_id: string;
    clinic_name: string;
}

export interface ClinicTypeEntity {
    clinic_type_id: string;
    clinic_type_name: string;
}